pyproject.toml
[build-system]
requires = ["hatchling>=1.24.2"]
build-backend = "hatchling.build"

[project]
name = "coffee-optimiser-backend"
version = "0.1.0"
description = "Production-grade backend platform for Coffee Optimiser"
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "Coffee Optimiser" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlalchemy>=2.0.32",
  "alembic>=1.13.2",
  "pydantic>=2.8.2",
  "pydantic-settings>=2.4.0",
  "python-jose[cryptography]>=3.3.0",
  "passlib[bcrypt]>=1.7.4",
  "psycopg[binary]>=3.2.1",
  "optuna>=3.6.1",
  "typer>=0.12.3",
  "httpx>=0.27.0",
  "python-multipart>=0.0.9",
  "orjson>=3.10.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
  "pytest-cov>=5.0.0",
  "mypy>=1.11.1",
  "ruff>=0.6.1",
  "black>=24.8.0",
  "pre-commit>=3.8.0",
]

[project.scripts]
coffee = "coffee_backend.cli.app:app"

[tool.hatch.build.targets.wheel]
packages = ["src/coffee_backend"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N", "C4", "SIM", "RUF"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]
exclude = ["src/coffee_backend/db/migrations"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
